+++ aplicacion/disponibilidad/__init__.py
@@ -1,0 +1,1 @@
+"""Inventario determinístico del hotel."""